refactor(objectives): move Evidence routing into analysis
Why:
- Transient route records lived apart from the service logic that selected and ordered Sources, leaving the confirmed-analysis coordinator responsible for the entire inspection queue.

What:
- Move Source candidate routing and document round-robin queue ordering to one direct analysis owner, and update the coordinator and tests to consume it without a compatibility path.

Impact:
- Model calls, prompt payloads, Source identities, route roles, extractability, confidence, progress, fallback behavior, and extraction order remain unchanged.

Verification:
- All 120 targeted Objective analysis tests passed; the new routing owner also passed compilation and lint, and the ownership and docs checks found no old routing path.
